move Bucket/File* into a folder
