+++ catalog/app/containers/Bucket/File/index.ts
@@ -1,0 +1,1 @@
+export { default } from './File'
